Extracted start methods into various modules to clean main up
