+++ nym-api/src/main.rs
@@ -4,48 +4,37 @@
 #[macro_use]
 extern crate rocket;
 
-use crate::network_monitor::NetworkMonitorBuilder;
-use crate::node_status_api::cache::refresher::NodeStatusCacheRefresher;
-use crate::node_status_api::uptime_updater::HistoricalUptimeUpdater;
-use crate::support::config::Config;
+use crate::support::cli;
 use crate::support::storage;
 use ::config::defaults::setup_env;
-use ::config::defaults::var_names::{MIXNET_CONTRACT_ADDRESS, MIX_DENOM};
-use ::config::{NymConfig, OptionalSet};
+use ::config::defaults::var_names::MIX_DENOM;
 use anyhow::Result;
 use build_information::BinaryBuildInformation;
 use clap::ArgMatches;
-use log::{info, warn};
-use logging::setup_logging;
-use node_status_api::NodeStatusCache;
-use nym_contract_cache::cache::refresher::NymContractCacheRefresher;
-use nym_contract_cache::cache::NymContractCache;
-use rocket::fairing::AdHoc;
-use rocket::http::Method;
-use rocket::{Ignite, Rocket};
-use rocket_cors::{AllowedHeaders, AllowedOrigins, Cors};
-use std::path::PathBuf;
-use std::str::FromStr;
-use std::sync::Arc;
-use task::{wait_for_signal, TaskManager};
-use tokio::sync::Notify;
-use validator_client::nyxd::{self, SigningNyxdClient};
-
-use crate::epoch_operations::RewardedSetUpdater;
-use crate::support::cli;
 #[cfg(feature = "coconut")]
 use coconut::{
     comm::QueryCommunicationChannel,
     dkg::controller::{init_keypair, DkgController},
     InternalSignRequest,
 };
+use log::info;
 use logging::setup_logging;
+use logging::setup_logging;
+use node_status_api::NodeStatusCache;
+use nym_contract_cache::cache::NymContractCache;
+use std::path::PathBuf;
+use std::str::FromStr;
+use std::sync::Arc;
+use support::{http, nyxd, process_runner};
+use task::{wait_for_signal, TaskManager};
+use tokio::sync::Notify;
+#[cfg(feature = "coconut")]
+use url::Url;
 #[cfg(feature = "coconut")]
 use rand::rngs::OsRng;
 use support::{http, nyxd, process_runner};
 use support::{nyxd, openapi};
 
-mod caching_support;
 mod circulating_supply_api;
 mod epoch_operations;
 mod network_monitor;
@@ -150,78 +139,26 @@ async fn wait_for_interrupt(mut shutdown: TaskManager) {
     log::info!("Sending shutdown");
     shutdown.signal_shutdown().ok();
 
-fn setup_cors() -> Result<Cors> {
-    let allowed_origins = AllowedOrigins::all();
+#[tokio::main]
+async fn main() -> Result<()> {
+    println!("Starting nym api...");
 
-    // You can also deserialize this
-    let cors = rocket_cors::CorsOptions {
-        allowed_origins,
-        allowed_methods: vec![Method::Post, Method::Get]
-            .into_iter()
-            .map(From::from)
-            .collect(),
-        allowed_headers: AllowedHeaders::all(),
-        allow_credentials: true,
-        ..Default::default()
-    }
-    .to_cors()?;
+    cfg_if::cfg_if! {if #[cfg(feature = "console-subscriber")] {
+        // instrument tokio console subscriber needs RUSTFLAGS="--cfg tokio_unstable" at build time
+        console_subscriber::init();
+    }}
 
-    Ok(cors)
+    setup_logging();
+    let args = cli::parse_args();
+    let config_env_file = args
+        .value_of(cli::args::CONFIG_ENV_FILE)
+        .map(|s| PathBuf::from_str(s).expect("invalid env config file"));
+    setup_env(config_env_file);
+    run_nym_api(args).await
 }
 
-fn setup_liftoff_notify(notify: Arc<Notify>) -> AdHoc {
-    AdHoc::on_liftoff("Liftoff notifier", |_| {
-        Box::pin(async move { notify.notify_one() })
-    })
-}
-
-fn setup_network_monitor<'a>(
-    config: &'a Config,
-    _nyxd_client: nyxd::Client<SigningNyxdClient>,
-    system_version: &str,
-    rocket: &Rocket<Ignite>,
-) -> Option<NetworkMonitorBuilder<'a>> {
-    if !config.get_network_monitor_enabled() {
-        return None;
-    }
-
-    // get instances of managed states
-    let node_status_storage = rocket.state::<storage::NymApiStorage>().unwrap().clone();
-    let nym_contract_cache = rocket.state::<NymContractCache>().unwrap().clone();
-
-    Some(NetworkMonitorBuilder::new(
-        config,
-        _nyxd_client,
-        system_version,
-        node_status_storage,
-        nym_contract_cache,
-    ))
-}
-
-// fn setup_circulating_supply() -> Option<>
-// }
-
-async fn run_nym_api(args: ApiArgs) -> Result<()> {
-    let system_version = env!("CARGO_PKG_VERSION");
-
-    // try to load config from the file, if it doesn't exist, use default values
-    let id = args.id.as_deref();
-    let (config, _already_inited) = match Config::load_from_file(id) {
-        Ok(cfg) => (cfg, true),
-        Err(_) => {
-            let config_path = Config::default_config_file_path(id)
-                .into_os_string()
-                .into_string()
-                .unwrap();
-            warn!(
-                "Could not load the configuration file from {config_path}. Either the file did not exist or was malformed. Using the default values instead",                
-            );
-            (Config::new(), false)
-        }
-    };
-
-    let save_to_file = args.save_config;
-    let config = cli::override_config(config, args);
+async fn run_nym_api(cli_args: ApiArgs) -> Result<()> {
+    let (system_version, config) = cli::build_config(cli_args);
 
     #[cfg(feature = "coconut")]
     if !_already_inited {
@@ -240,10 +177,9 @@ async fn run_nym_api(args: ApiArgs) -> Result<()> {
     }
 
     let mix_denom = std::env::var(MIX_DENOM).expect("mix denom not set");
-
     let signing_nyxd_client = nyxd::Client::new_signing(&config);
-
     let liftoff_notify = Arc::new(Notify::new());
+
     // We need a bigger timeout
     let shutdown = TaskManager::new(10);
 
@@ -260,15 +196,9 @@ async fn run_nym_api(args: ApiArgs) -> Result<()> {
         coconut_keypair.clone(),
     )
     .await?;
-    let monitor_builder = setup_network_monitor(
-        &config,
-        signing_nyxd_client.clone(),
-        system_version,
-        &rocket,
-    );
 
-    let nym_contract_cache = rocket.state::<NymContractCache>().unwrap().clone();
-    let node_status_cache = rocket.state::<NodeStatusCache>().unwrap().clone();
+    let nym_contract_cache_state = rocket.state::<NymContractCache>().unwrap().clone();
+    let node_status_cache_state = rocket.state::<NodeStatusCache>().unwrap().clone();
 
     #[cfg(feature = "coconut")]
     {
@@ -282,70 +212,40 @@ async fn run_nym_api(args: ApiArgs) -> Result<()> {
     // if network monitor is disabled, we're not going to be sending any rewarding hence
     // we're not starting signing client
     let nym_contract_cache_listener = if config.get_network_monitor_enabled() {
-        // Main storage
-        let storage = rocket.state::<storage::NymApiStorage>().unwrap().clone();
-
-        // setup our daily uptime updater. Note that if network monitor is disabled, then we have
-        // no data for the updates and hence we don't need to start it up
-        let uptime_updater = HistoricalUptimeUpdater::new(storage.clone());
-        let shutdown_listener = shutdown.subscribe();
-        tokio::spawn(async move { uptime_updater.run(shutdown_listener).await });
-
-        // spawn the nym contract cache refresher
-        let nym_contract_cache_refresher = NymContractCacheRefresher::new(
-            signing_nyxd_client.clone(),
-            config.get_caching_interval(),
-            nym_contract_cache.clone(),
-        );
-        let nym_contract_cache_listener = nym_contract_cache_refresher.subscribe();
-        let shutdown_listener = shutdown.subscribe();
-        tokio::spawn(async move { nym_contract_cache_refresher.run(shutdown_listener).await });
-
-        // spawn rewarded set updater
-        if config.get_rewarding_enabled() {
-            let mut rewarded_set_updater =
-                RewardedSetUpdater::new(signing_nyxd_client, nym_contract_cache.clone(), storage)
-                    .await?;
-            let shutdown_listener = shutdown.subscribe();
-            tokio::spawn(async move { rewarded_set_updater.run(shutdown_listener).await.unwrap() });
-        }
-        nym_contract_cache_listener
+        nym_contract_cache::start_with_signing(
+            &rocket,
+            &shutdown,
+            &signing_nyxd_client,
+            &config,
+            &nym_contract_cache_state,
+        )
+        .await?
     } else {
-        // Spawn the nym contract cache refresher.
-        // When the network monitor is not enabled, we spawn the nym contract cache refresher task
-        // with just a nyxd client, in contrast to a signing client.
-        let nyxd_client = nyxd::Client::new_query(&config);
-        let nym_contract_cache_refresher = NymContractCacheRefresher::new(
-            nyxd_client,
-            config.get_caching_interval(),
-            nym_contract_cache.clone(),
-        );
-        let nym_contract_cache_listener = nym_contract_cache_refresher.subscribe();
-        let shutdown_listener = shutdown.subscribe();
-        tokio::spawn(async move { nym_contract_cache_refresher.run(shutdown_listener).await });
-
-        nym_contract_cache_listener
+        nym_contract_cache::start_without_signing(&config, &nym_contract_cache_state, &shutdown)
     };
 
-    // Spawn the node status cache refresher.
-    // It is primarily refreshed in-sync with the nym contract cache, however provide a fallback
-    // caching interval that is twice the nym contract cache
-    let storage = rocket.state::<storage::NymApiStorage>().cloned();
-    let mut nym_api_cache_refresher = NodeStatusCacheRefresher::new(
-        node_status_cache,
-        config.get_caching_interval().saturating_mul(2),
-        nym_contract_cache,
+    node_status_api::start_cache_refresh(
+        &rocket,
+        node_status_cache_state,
+        &config,
+        nym_contract_cache_state,
         nym_contract_cache_listener,
-        storage,
+        &shutdown,
     );
-    let shutdown_listener = shutdown.subscribe();
-    tokio::spawn(async move { nym_api_cache_refresher.run(shutdown_listener).await });
 
-    // launch the rocket!
-    // Rocket handles shutdown on it's own, but its shutdown handling should be incorporated
-    // with that of the rest of the tasks.
-    // Currently it's runtime is forcefully terminated once the nym-api exits.
+    let monitor_builder = network_monitor::setup(
+        &config,
+        signing_nyxd_client.clone(),
+        system_version,
+        &rocket,
+    );
+
+    // Rocket handles shutdown on its own, but its shutdown handling should be incorporated
+    // with that of the rest of the tasks. Currently its runtime is forcefully terminated once
+    // nym-api exits.
     let shutdown_handle = rocket.shutdown();
+
+    // Launch the rocket!
     tokio::spawn(rocket.launch());
 
     // to finish building our monitor, we need to have rocket up and running so that we could
@@ -366,19 +266,4 @@ async fn run_nym_api(args: ApiArgs) -> Result<()> {
     shutdown_handle.notify();
 
     Ok(())
-}
-
-#[tokio::main]
-async fn main() -> Result<()> {
-    println!("Starting nym api...");
-
-    cfg_if::cfg_if! {if #[cfg(feature = "console-subscriber")] {
-        // instrument tokio console subscriber needs RUSTFLAGS="--cfg tokio_unstable" at build time
-        console_subscriber::init();
-    }}
-
-    setup_logging();
-    let args = ApiArgs::parse();
-    setup_env(args.config_env_file.as_ref());
-    run_nym_api(args).await
 }
