+++ nym-api/src/support/http/mod.rs
@@ -4,7 +4,11 @@ use crate::nym_contract_cache::cache::NymContractCache;
 use crate::support::config::Config;
 use crate::support::{nyxd, storage};
 use crate::{circulating_supply_api, nym_contract_cache};
+use anyhow::Result;
+use rocket::fairing::AdHoc;
+use rocket::http::Method;
 use rocket::{Ignite, Rocket};
+use rocket_cors::{AllowedHeaders, AllowedOrigins, Cors};
 use rocket_okapi::mount_endpoints_and_merged_docs;
 use rocket_okapi::swagger_ui::make_swagger_ui;
 use std::sync::Arc;
@@ -43,8 +47,8 @@ pub(crate) async fn setup_rocket(
 
     let rocket = rocket
         .mount("/swagger", make_swagger_ui(&openapi::get_docs()))
-        .attach(crate::setup_cors()?)
-        .attach(crate::setup_liftoff_notify(liftoff_notify))
+        .attach(setup_cors()?)
+        .attach(setup_liftoff_notify(liftoff_notify))
         .attach(NymContractCache::stage())
         .attach(NodeStatusCache::stage())
         .attach(CirculatingSupplyCache::stage());
@@ -78,4 +82,29 @@ pub(crate) async fn setup_rocket(
     };
 
     Ok(rocket.ignite().await?)
+}
+
+fn setup_cors() -> Result<Cors> {
+    let allowed_origins = AllowedOrigins::all();
+
+    // You can also deserialize this
+    let cors = rocket_cors::CorsOptions {
+        allowed_origins,
+        allowed_methods: vec![Method::Post, Method::Get]
+            .into_iter()
+            .map(From::from)
+            .collect(),
+        allowed_headers: AllowedHeaders::all(),
+        allow_credentials: true,
+        ..Default::default()
+    }
+    .to_cors()?;
+
+    Ok(cors)
+}
+
+fn setup_liftoff_notify(notify: Arc<Notify>) -> AdHoc {
+    AdHoc::on_liftoff("Liftoff notifier", |_| {
+        Box::pin(async move { notify.notify_one() })
+    })
 }
